initial steps on improving the code
